Consonance Functions for Correlations

Computes consonance intervals to produce P- and S-value functions for correlational analysesusing the cor.test function in base R and places the interval limits for each interval levelinto a data frame along with the corresponding p-values and s-values.

curve_corr(x, y, alternative, method, steps = 10000, table = TRUE)

Arguments

x
A vector that contains the data for one of the variables that will be analyzed for correlational analysis.

y
A vector that contains the data for one of the variables that will be analyzed for correlational analysis.

alternative
Indicates the alternative hypothesis and must be one of "two.sided", "greater" or "less". You can specify just the initial letter. "greater" corresponds to positive association, "less" to negative association.

method
A character string indicating which correlation coefficient is to be used for the test. One of "pearson", "kendall", or "spearman", can be abbreviated.

steps
Indicates how many consonance intervals are to be calculated at various levels. For example, setting this to 100 will produce 100 consonance intervals from 0 to 100. Setting this to 10000 will produce more consonance levels. By default, it is set to 1000. Increasing the number substantially is not recommended as it will take longer to produce all the intervals and store them into a dataframe.

table
Indicates whether or not a table output with some relevant statistics should be generated. The default is TRUE and generates a table which is included in the list object.

Value

A list with 3 items where the dataframe of values is in the first object, the values needed to calculate the density function in the second, and the table for the values in the third if table = TRUE.

Examples

GroupA <- rnorm(50)
GroupB <- rnorm(50)
joe <- curve_corr(x = GroupA, y = GroupB, alternative = "two.sided", method = "pearson")
tibble::tibble(joe[[1]])

#> # A tibble: 10,000 x 1
#>    `joe[[1]]`$lowe… $upper.limit $intrvl.width $intrvl.level  $cdf $pvalue
#>               <dbl>        <dbl>         <dbl>         <dbl> <dbl>   <dbl>
#>  1            0.106        0.106     0              0        0.5     1    
#>  2            0.106        0.106     0.0000362      0.0001   0.500   1.000
#>  3            0.106        0.106     0.0000723      0.0002   0.500   1.000
#>  4            0.106        0.106     0.000108       0.000300 0.500   1.000
#>  5            0.106        0.106     0.000145       0.0004   0.500   1.000
#>  6            0.106        0.106     0.000181       0.0005   0.500   1.000
#>  7            0.106        0.106     0.000217       0.000600 0.500   0.999
#>  8            0.106        0.106     0.000253       0.0007   0.500   0.999
#>  9            0.106        0.106     0.000289       0.0008   0.500   0.999
#> 10            0.106        0.106     0.000325       0.0009   0.500   0.999
#> # … with 9,990 more rows, and 1 more variable: $svalue <dbl>