Compute Profile Likelihood Functions

Compute Profile Likelihood Functions

curve_lik(likobject, data, table = TRUE)

Arguments

likobject
An object from the ProfileLikelihood package

data
The dataframe that was used to create the likelihood object in the ProfileLikelihood package.

table
Indicates whether or not a table output with some relevant statistics should be generated. The default is TRUE and generates a table which is included in the list object.

Value

A list with 2 items where the dataframe of values is in the first object, and the table for the values in the second if table = TRUE.

Examples

library(ProfileLikelihood)

#> Loading required package: nlme

#> Loading required package: MASS

data(dataglm)
xx <- profilelike.glm(y ~ x1 + x2, dataglm, profile.theta = "group", binomial("logit"))

#> Warning message: provide lo.theta and hi.theta 

lik <- curve_lik(xx, dataglm)
tibble::tibble(lik[[1]])

#> # A tibble: 300 x 1
#>    `lik[[1]]`$values $likelihood $loglikelihood  $support $deviancestat
#>                <dbl>       <dbl>          <dbl>     <dbl>         <dbl>
#>  1             -1.41    9.26e-21          -9.79 0.0000560          9.79
#>  2             -1.39    1.05e-20          -9.66 0.0000638          9.66
#>  3             -1.37    1.20e-20          -9.53 0.0000726          9.53
#>  4             -1.35    1.37e-20          -9.40 0.0000826          9.40
#>  5             -1.34    1.55e-20          -9.27 0.0000940          9.27
#>  6             -1.32    1.76e-20          -9.15 0.000107           9.15
#>  7             -1.30    2.00e-20          -9.02 0.000121           9.02
#>  8             -1.28    2.27e-20          -8.89 0.000137           8.89
#>  9             -1.26    2.57e-20          -8.77 0.000156           8.77
#> 10             -1.24    2.92e-20          -8.64 0.000176           8.64
#> # … with 290 more rows