Reverse Engineer Consonance / Likelihood Functions Using the Point Estimate and Confidence Limits

Using the confidence limits and point estimates from a dataset, one can use these estimates to compute thousands of consonance intervals and graph the intervals to form a consonance and surprisal function. The intervals are calculated from the approximated normal distribution.

curve_rev(point, LL, UL, type = "c", measure = "default", steps = 10000,
  table = TRUE)

Arguments

point
The point estimate from an analysis. Ex: 1.20

LL
The lower confidence limit from an analysis Ex: 1.0

UL
The upper confidence limit from an analysis Ex: 1.4

type
Indicates whether the produced result should be a consonance function or a likelihood function. The default is "c" for consonance and likelihood can be set via "l".

measure
The type of data being used. If they involve mean differences, then the "default" option should be used, which is also the default setting. If the data are ratios, then the "ratio" option should be used.

steps
Indicates how many consonance intervals are to be calculated at various levels. For example, setting this to 100 will produce 100 consonance intervals from 0 to 100. Setting this to 10000 will produce more consonance levels. By default, it is set to 1000. Increasing the number substantially is not recommended as it will take longer to produce all the intervals and store them into a dataframe.

table
Indicates whether or not a table output with some relevant statistics should be generated. The default is TRUE and generates a table which is included in the list object.

Value

A list with 3 items where the dataframe of values is in the first object, the values needed to calculate the density function in the second, and the table for the values in the third if table = TRUE.

See also

Examples

# From a real published study. Point estimate of the result was hazard ratio of 1.61 and
# lower bound of the interval is 0.997 while upper bound of the interval is 2.59.
#
df <- curve_rev(point = 1.61, LL = 0.997, UL = 2.59, measure = "ratio")

tibble::tibble(df[[1]])

#> # A tibble: 9,999 x 1
#>    `df[[1]]`$lower… $upper.limit $intrvl.width $intrvl.level  $cdf $pvalue
#>               <dbl>        <dbl>         <dbl>         <dbl> <dbl>   <dbl>
#>  1            0.624         4.15          3.53         1.00   1.00 1.00e-4
#>  2            0.651         3.98          3.33         1.00   1.00 2.00e-4
#>  3            0.667         3.88          3.22         1.00   1.00 3.00e-4
#>  4            0.680         3.81          3.13         1.00   1.00 4.00e-4
#>  5            0.690         3.76          3.07         1.00   1.00 5.00e-4
#>  6            0.698         3.71          3.02         0.999  1.00 6.00e-4
#>  7            0.705         3.68          2.97         0.999  1.00 7.00e-4
#>  8            0.712         3.64          2.93         0.999  1.00 8.00e-4
#>  9            0.717         3.61          2.90         0.999  1.00 9.00e-4
#> 10            0.722         3.59          2.87         0.999  1.00 1.00e-3
#> # … with 9,989 more rows, and 1 more variable: $svalue <dbl>