Produces a Table of Relevant Statistics from the `concurve` functions

curve_table(data, levels, type = "data.frame")

Arguments

data

levels

type

Details

Value

References

Note

See also

Examples

library(concurve)

GroupA <- rnorm(500)
GroupB <- rnorm(500)

RandomData <- data.frame(GroupA, GroupB)

intervalsdf <- curve_mean(GroupA, GroupB,
                          data = RandomData, method = "default"
)

(z <- curve_table(intervalsdf, type = "data.frame"))

#>      Lower Limit Upper Limit Interval Width Interval Level (%) P-value
#> 2501       0.043       0.083          0.040               25.0   0.750
#> 5001       0.021       0.105          0.084               50.0   0.500
#> 7501      -0.008       0.135          0.143               75.0   0.250
#> 8001      -0.016       0.143          0.159               80.0   0.200
#> 8501      -0.026       0.153          0.179               85.0   0.150
#> 9001      -0.039       0.166          0.205               90.0   0.100
#> 9501      -0.059       0.185          0.244               95.0   0.050
#> 9751      -0.076       0.203          0.279               97.5   0.025
#> 9901      -0.097       0.224          0.321               99.0   0.010
#>      S-value (bits)
#> 2501          0.415
#> 5001          1.000
#> 7501          2.000
#> 8001          2.322
#> 8501          2.737
#> 9001          3.322
#> 9501          4.322
#> 9751          5.322
#> 9901          6.644

(z <- curve_table(intervalsdf, type = "tibble"))

#> # A tibble: 9 x 1
#>   subdf$`Lower Li… $`Upper Limit` $`Interval Widt… $`Interval Leve… $`P-value`
#>              <dbl>          <dbl>            <dbl>            <dbl>      <dbl>
#> 1            0.043          0.083            0.04              25        0.75 
#> 2            0.021          0.105            0.084             50        0.5  
#> 3           -0.008          0.135            0.143             75        0.25 
#> 4           -0.016          0.143            0.159             80        0.2  
#> 5           -0.026          0.153            0.179             85        0.15 
#> 6           -0.039          0.166            0.205             90        0.1  
#> 7           -0.059          0.185            0.244             95        0.05 
#> 8           -0.076          0.203            0.279             97.5      0.025
#> 9           -0.097          0.224            0.321             99        0.01 
#> # … with 1 more variable: $`S-value (bits)` <dbl>

(z <- curve_table(intervalsdf, type = "latex"))

#> 
#> 
#> |     | Lower Limit| Upper Limit| Interval Width| Interval Level (%)| P-value| S-value (bits)|
#> |:----|-----------:|-----------:|--------------:|------------------:|-------:|--------------:|
#> |2501 |       0.043|       0.083|          0.040|               25.0|   0.750|          0.415|
#> |5001 |       0.021|       0.105|          0.084|               50.0|   0.500|          1.000|
#> |7501 |      -0.008|       0.135|          0.143|               75.0|   0.250|          2.000|
#> |8001 |      -0.016|       0.143|          0.159|               80.0|   0.200|          2.322|
#> |8501 |      -0.026|       0.153|          0.179|               85.0|   0.150|          2.737|
#> |9001 |      -0.039|       0.166|          0.205|               90.0|   0.100|          3.322|
#> |9501 |      -0.059|       0.185|          0.244|               95.0|   0.050|          4.322|
#> |9751 |      -0.076|       0.203|          0.279|               97.5|   0.025|          5.322|
#> |9901 |      -0.097|       0.224|          0.321|               99.0|   0.010|          6.644|

(z <- curve_table(intervalsdf, type = "image"))

#> a flextable object.
#> col_keys: `Lower Limit`, `Upper Limit`, `Interval Width`, `Interval Level (%)`, `P-value`, `S-value (bits)` 
#> header has 1 row(s) 
#> body has 9 row(s) 
#> original dataset sample: 
#>      Lower Limit Upper Limit Interval Width Interval Level (%) P-value
#> 2501       0.043       0.083          0.040                 25    0.75
#> 5001       0.021       0.105          0.084                 50    0.50
#> 7501      -0.008       0.135          0.143                 75    0.25
#> 8001      -0.016       0.143          0.159                 80    0.20
#> 8501      -0.026       0.153          0.179                 85    0.15
#>      S-value (bits)
#> 2501          0.415
#> 5001          1.000
#> 7501          2.000
#> 8001          2.322
#> 8501          2.737