Computes consonance intervals for correlations

Computes consonance intervals to produce P- and S-value functions for correlational analyses using the cor.test function in base R and places the interval limits for each interval level into a data frame along with the corresponding p-values and s-values.

curve_corr(x, y, alternative, method, steps = 10000)

Arguments

x
A vector that contains the data for one of the variables that will be analyzed for correlational analysis.

y
A vector that contains the data for one of the variables that will be analyzed for correlational analysis.

alternative
Indicates the alternative hypothesis and must be one of "two.sided", "greater" or "less". You can specify just the initial letter. "greater" corresponds to positive association, "less" to negative association.

method
A character string indicating which correlation coefficient is to be used for the test. One of "pearson", "kendall", or "spearman", can be abbreviated.

steps
Indicates how many consonance intervals are to be calculated at various levels. For example, setting this to 100 will produce 100 consonance intervals from 0 to 100. Setting this to 10000 will produce more consonance levels. By default, it is set to 1000. Increasing the number substantially is not recommended as it will take longer to produce all the intervals and store them into a dataframe.

References

Poole C. Beyond the confidence interval. Am J Public Health. 1987;77(2):195-199.

Sullivan KM, Foster DA. Use of the confidence interval function. Epidemiology. 1990;1(1):39-42.

Rothman KJ, Greenland S, Lash TL, Others. Modern epidemiology. 2008.

Examples

GroupA <- rnorm(50)
GroupB <- rnorm(50)

joe <- curve_corr(x = GroupA, y = GroupB,
 alternative = "two.sided", method = "pearson")

tibble::tibble(joe)

#> # A tibble: 10,000 x 1
#>    joe$lower.limit $upper.limit $limit.ratio $intrvl.level $pvalue  $svalue
#>              <dbl>        <dbl>        <dbl>         <dbl>   <dbl>    <dbl>
#>  1           0.398        0.398         1         0          1     0       
#>  2           0.398        0.398         1.00      0.0001     1.000 0.000144
#>  3           0.398        0.398         1.00      0.0002     1.000 0.000289
#>  4           0.398        0.398         1.00      0.000300   1.000 0.000433
#>  5           0.398        0.398         1.00      0.0004     1.000 0.000577
#>  6           0.398        0.398         1.00      0.0005     1.000 0.000722
#>  7           0.398        0.398         1.00      0.000600   0.999 0.000866
#>  8           0.398        0.398         1.00      0.0007     0.999 0.00101 
#>  9           0.398        0.398         1.00      0.0008     0.999 0.00115 
#> 10           0.398        0.398         1.00      0.0009     0.999 0.00130 
#> # … with 9,990 more rows