Consonance Functions For Linear Models.

Computes thousands of consonance (confidence) intervals for the chosen parameter in the selected model (ANOVA, ANCOVA, regression, logistic regression) and places the interval limits for each interval level into a data frame along with the corresponding p-values and s-values.

curve_gen(model, var, method = "wald", steps = 1000, table = TRUE)

Arguments

model
The statistical model of interest (ANOVA, regression, logistic regression) is to be indicated here.

var
The variable of interest from the model (coefficients, intercept) for which the intervals are to be produced.

method
Chooses the method to be used to calculate the consonance intervals. There are currently four methods: "default", "wald", "lm", and "boot". The "default" method uses the profile likelihood method to compute intervals and can be used for models created by the 'lm' function. The "wald" method is typicallywhat most people are familiar with when computing intervals based on the calculated standard error. The "lm" method allows this function to be used for specific scenarios like logistic regression and the 'glm' function. The "boot" method allows for bootstrapping at certain levels.

steps
Indicates how many consonance intervals are to be calculated at various levels. For example, setting this to 100 will produce 100 consonance intervals from 0 to 100. Setting this to 10000 will produce more consonance levels. By default, it is set to 1000. Increasing the number substantially is not recommended as it will take longer to produce all the intervals and store them into a dataframe.

table
Indicates whether or not a table output with some relevant statistics should be generated. The default is TRUE and generates a table which is included in the list object.

Value

A list with 3 items where the dataframe of values is in the first object, the values needed to calculate the density function in the second, and the table for the values in the third if table = TRUE.

Examples

# \donttest{
# Simulate random data
GroupA <- rnorm(50)
GroupB <- rnorm(50)
RandomData <- data.frame(GroupA, GroupB)
rob <- lm(GroupA ~ GroupB, data = RandomData)
bob <- curve_gen(rob, "GroupB")
tibble::tibble(bob[[1]])

#> # A tibble: 998 x 1
#>    `bob[[1]]`$lowe… $upper.limit $intrvl.width $intrvl.level  $cdf $pvalue
#>               <dbl>        <dbl>         <dbl>         <dbl> <dbl>   <dbl>
#>  1         -0.00612     -0.00570      0.000421         0.001 0.500   0.999
#>  2         -0.00633     -0.00549      0.000841         0.002 0.501   0.998
#>  3         -0.00654     -0.00528      0.00126          0.003 0.501   0.997
#>  4         -0.00675     -0.00507      0.00168          0.004 0.502   0.996
#>  5         -0.00696     -0.00486      0.00210          0.005 0.502   0.995
#>  6         -0.00717     -0.00465      0.00252          0.006 0.503   0.994
#>  7         -0.00738     -0.00444      0.00294          0.007 0.503   0.993
#>  8         -0.00759     -0.00423      0.00337          0.008 0.504   0.992
#>  9         -0.00780     -0.00402      0.00379          0.009 0.504   0.991
#> 10         -0.00801     -0.00381      0.00421          0.01  0.505   0.99 
#> # … with 988 more rows, and 1 more variable: $svalue <dbl>

# }