Consonance Functions For Linear Mixed-Effects Models.

Computes thousands of consonance (confidence) intervals for the chosen parameter in the selected lme4 model and places the interval limits for each interval level into a data frame along with the corresponding p-values and s-values.

curve_lmer(object, parm, method = "profile", zeta = NULL, nsim = NULL,
  FUN = NULL, boot.type = NULL, steps = 1000, mc.cores = 1,
  table = FALSE)

Arguments

object
The statistical model of interest from lme4 is to be indicated here.

parm
The variable of interest from the model (coefficients, intercept) for which the intervals are to be produced.

method
Chooses the method to be used to calculate the consonance intervals. There are currently four methods: "default", "wald", "lm", and "boot". The "default" method uses the profile likelihood method to compute intervals and can be used for models created by the 'lm' function. The "wald" method is typicallywhat most people are familiar with when computing intervals based on the calculated standard error. The "lm" method allows this function to be used for specific scenarios like logistic regression and the 'glm' function. The "boot" method allows for bootstrapping at certain levels.

zeta
(for method = "profile" only:) likelihood cutoff (if not specified, as by default, computed from level).

nsim
number of simulations for parametric bootstrap intervals.

FUN
function; if NULL, an internal function that returns the fixed-effect parameters as well as the random-effect parameters on the standard deviation/correlationscale will be used.

boot.type
bootstrap confidence interval type, as described in boot.ci. (Methods ‘stud’ and ‘bca’ are unavailable because they require additional components to be calculated.)

steps
Indicates how many consonance intervals are to be calculated at various levels. For example, setting this to 100 will produce 100 consonance intervals from 0 to 100. Setting this to 10000 will produce more consonance levels. By default, it is set to 1000. Increasing the number substantially is not recommended as it will take longer to produce all the intervals and store them into a dataframe.

mc.cores
For parallel processing. Defaults to 1 core.

table
Indicates whether or not a table output with some relevant statistics should be generated. The default is TRUE and generates a table which is included in the list object.

Value

A list with 3 items where the dataframe of values is in the first object, the values needed to calculate the density function in the second, and the table for the values in the third if table = TRUE.

Examples

# \donttest{
library(lme4)

#> Loading required package: Matrix

#> 
#> Attaching package: ‘lme4’

#> The following object is masked from ‘package:nlme’:
#> 
#>     lmList

m <- ~ x + (x || g)
expandDoubleVerts(m)

#> ~x + ((1 | g) + (0 + x | g))
#> <environment: 0x7fac72fd5110>

set.seed(101)
dd <- expand.grid(f=factor(letters[1:3]),g=factor(1:200),rep=1:3)
dd$y <- simulate(~f + (1|g) + (0+dummy(f,"b")|g) + (0+dummy(f,"c")|g),
                newdata=dd,
                newparams=list(beta=rep(0,3),
                               theta=c(1,2,1),
                               sigma=1),
                family=gaussian)[[1]]

#> theta parameter vector not named: assuming same order as internal vector

#> beta parameter vector not named: assuming same order as internal vector

m1 <- lmer(y~f+(f|g),data=dd)
VarCorr(m1)

#>  Groups   Name        Std.Dev. Corr         
#>  g        (Intercept) 0.95684               
#>           fb          1.97300   0.106       
#>           fc          0.96426   0.109 -0.086
#>  Residual             1.02172               

m2 <- lmer(y~f+(1|g) + (0+dummy(f,"b")|g) + (0+dummy(f,"c")|g),
          data=dd)
VarCorr(m2)

#>  Groups   Name          Std.Dev.
#>  g        (Intercept)   0.98657 
#>  g.1      dummy(f, "b") 2.00643 
#>  g.2      dummy(f, "c") 0.99618 
#>  Residual               1.01770 

confint.merMod(m2, parm = "fb", level = 0.95)

#> Computing profile confidence intervals ...

#>         2.5 %    97.5 %
#> fb -0.3041212 0.2991399

z <- curve_lmer(object = m2, parm = "fb", steps = 1000, mc.cores = 8, table = TRUE)
# }