Meta-analytic Consonance Function

Computes thousands of consonance (confidence) intervals for the chosen parameter in the meta-analysis done by the metafor package and places the interval limits for each interval level into a data frame along with the corresponding p-values and s-values.

curve_meta(x, measure = "default", steps = 10000, table = TRUE)

Arguments

x
Object where the meta-analysis parameters are stored, typically a list produced by 'metafor'

measure
Indicates whether the object has a log transformation or is normal/default. The default setting is "default. If the measure is set to "ratio", it will take logarithmically transformed values and convert them back to normal values in the dataframe. This is typically a setting used for binary outcomes such as risk ratios, hazard ratios, and odds ratios.

steps
Indicates how many consonance intervals are to be calculated at various levels. For example, setting this to 100 will produce 100 consonance intervals from 0 to 100. Setting this to 10000 will produce more consonance levels. By default, it is set to 1000. Increasing the number substantially is not recommended as it will take longer to produce all the intervals and store them into a dataframe.

table
Indicates whether or not a table output with some relevant statistics should be generated. The default is TRUE and generates a table which is included in the list object.

Examples

# Simulate random data for two groups in two studies
GroupAData <- runif(20, min = 0, max = 100)
GroupAMean <- round(mean(GroupAData), digits = 2)
GroupASD <- round(sd(GroupAData), digits = 2)

GroupBData <- runif(20, min = 0, max = 100)
GroupBMean <- round(mean(GroupBData), digits = 2)
GroupBSD <- round(sd(GroupBData), digits = 2)

GroupCData <- runif(20, min = 0, max = 100)
GroupCMean <- round(mean(GroupCData), digits = 2)
GroupCSD <- round(sd(GroupCData), digits = 2)

GroupDData <- runif(20, min = 0, max = 100)
GroupDMean <- round(mean(GroupDData), digits = 2)
GroupDSD <- round(sd(GroupDData), digits = 2)

# Combine the data

StudyName <- c("Study1", "Study2")
MeanTreatment <- c(GroupAMean, GroupCMean)
MeanControl <- c(GroupBMean, GroupDMean)
SDTreatment <- c(GroupASD, GroupCSD)
SDControl <- c(GroupBSD, GroupDSD)
NTreatment <- c(20, 20)
NControl <- c(20, 20)

metadf <- data.frame(
  StudyName, MeanTreatment, MeanControl,
  SDTreatment, SDControl, NTreatment, NControl
)

# Use metafor to calculate the standardized mean difference

library(metafor)

#> Loading required package: Matrix

#> Loading 'metafor' package (version 2.1-0). For an overview 
#> and introduction to the package please type: help(metafor).

dat <- escalc(
  measure = "SMD", m1i = MeanTreatment, sd1i = SDTreatment,
  n1i = NTreatment, m2i = MeanControl, sd2i = SDControl,
  n2i = NControl, data = metadf
)

# Pool the data using a particular method. Here "FE" is the fixed-effects model

res <- rma(yi, vi,
  data = dat, slab = paste(StudyName, sep = ", "),
  method = "FE", digits = 2
)

# Calculate the intervals using the metainterval function

metaf <- curve_meta(res)

tibble::tibble(metaf[[1]])

#> # A tibble: 10,000 x 1
#>    `metaf[[1]]`$lo… $upper.limit $intrvl.width $intrvl.level  $cdf $pvalue
#>               <dbl>        <dbl>         <dbl>         <dbl> <dbl>   <dbl>
#>  1           0.0617       0.0617     0              0        0.5     1    
#>  2           0.0616       0.0617     0.0000562      0.0001   0.500   1.000
#>  3           0.0616       0.0617     0.000112       0.0002   0.500   1.000
#>  4           0.0616       0.0617     0.000169       0.000300 0.500   1.000
#>  5           0.0615       0.0618     0.000225       0.0004   0.500   1.000
#>  6           0.0615       0.0618     0.000281       0.0005   0.500   1.000
#>  7           0.0615       0.0618     0.000337       0.000600 0.500   0.999
#>  8           0.0615       0.0619     0.000394       0.0007   0.500   0.999
#>  9           0.0614       0.0619     0.000450       0.0008   0.500   0.999
#> 10           0.0614       0.0619     0.000506       0.0009   0.500   0.999
#> # … with 9,990 more rows, and 1 more variable: $svalue <dbl>