Reverse engineer consonance and surprisal functions from confidence limits and point estimates

Using the confidence limits and point estimates from a dataset, one can use these estimates to compute thousands of consonance intervals and graph the intervals to form a consonance and surprisal function.

curve_rev(point, LL, UL, measure = "default")

Arguments

point
The point estimate from an analysis. Ex: 1.20

LL
The lower confidence limit from an analysis Ex: 1.0

UL
The upper confidence limit from an analysis Ex: 1.4

measure
The type of data being used. If they involve mean differences, then the "default" option should be used, which is also the default setting. If the data are ratios, then the "ratio" option should be used.

References

Poole C. Beyond the confidence interval. Am J Public Health. 1987;77(2):195-199.

Sullivan KM, Foster DA. Use of the confidence interval function. Epidemiology. 1990;1(1):39-42.

Rothman KJ, Greenland S, Lash TL, Others. Modern epidemiology. 2008.

Examples

# From a real published study. Point estimate of the result was hazard ratio of 1.61 and
# lower bound of the interval is 0.997 while upper bound of the interval is 2.59.

df <- curve_rev(point = 1.61, LL = 0.997, UL = 2.59, measure = "ratio")

tibble::tibble(df)

#> # A tibble: 9,999 x 1
#>    df$lower.limit $upper.limit $intrvl.level   $pvalue $svalue
#>             <dbl>        <dbl>         <dbl>     <dbl>   <dbl>
#>  1          0.624         4.15         1.000 0.0001000   13.3 
#>  2          0.651         3.98         1.000 0.000200    12.3 
#>  3          0.667         3.88         1.000 0.000300    11.7 
#>  4          0.680         3.81         1.000 0.000400    11.3 
#>  5          0.690         3.76         1.000 0.000500    11.0 
#>  6          0.698         3.71         0.999 0.0006      10.7 
#>  7          0.705         3.68         0.999 0.0007      10.5 
#>  8          0.712         3.64         0.999 0.0008      10.3 
#>  9          0.717         3.61         0.999 0.0009      10.1 
#> 10          0.722         3.59         0.999 0.001        9.97
#> # … with 9,989 more rows