Produce Tables For concurve Functions

Produces publication-ready tables with relevant statistics of interest for functions produced from the concurve package.

curve_table(data, levels, type = "c", format = "data.frame")

Arguments

data
Dataframe from a concurve function to produce a table for

levels
Levels of the consonance intervals or likelihood intervals that should be included in the table.

type
Indicates whether the table is for a consonance function or likelihood function. The default is set to "c" for consonance and can be switched to "l" for likelihood.

format
The format of the tables. The options include "data.frame" which is the default, "tibble", "docx" (which creates a table for a word document), "pptx" (which creates a table for powerpoint), "latex", (which creates a table for a TeX document), and "image", which produces an image of the table.

Examples

library(concurve)

GroupA <- rnorm(500)
GroupB <- rnorm(500)

RandomData <- data.frame(GroupA, GroupB)

intervalsdf <- curve_mean(GroupA, GroupB, data = RandomData, method = "default")

(z <- curve_table(intervalsdf[[1]], format = "data.frame"))

#>      Lower Limit Upper Limit Interval Width Interval Level (%)   CDF P-value
#> 2501       0.131       0.172          0.041               25.0 0.625   0.750
#> 5001       0.108       0.195          0.088               50.0 0.750   0.500
#> 7501       0.077       0.226          0.149               75.0 0.875   0.250
#> 8001       0.068       0.235          0.166               80.0 0.900   0.200
#> 8501       0.058       0.245          0.187               85.0 0.925   0.150
#> 9001       0.045       0.258          0.214               90.0 0.950   0.100
#> 9501       0.024       0.279          0.255               95.0 0.975   0.050
#> 9751       0.006       0.297          0.291               97.5 0.988   0.025
#> 9901      -0.016       0.319          0.335               99.0 0.995   0.010
#>      S-value (bits)
#> 2501          0.415
#> 5001          1.000
#> 7501          2.000
#> 8001          2.322
#> 8501          2.737
#> 9001          3.322
#> 9501          4.322
#> 9751          5.322
#> 9901          6.644

(z <- curve_table(intervalsdf[[1]], format = "tibble"))

#> # A tibble: 9 x 1
#>   subdf$`Lower Li… $`Upper Limit` $`Interval Widt… $`Interval Leve…  $CDF
#>              <dbl>          <dbl>            <dbl>            <dbl> <dbl>
#> 1            0.131          0.172            0.041             25   0.625
#> 2            0.108          0.195            0.088             50   0.75 
#> 3            0.077          0.226            0.149             75   0.875
#> 4            0.068          0.235            0.166             80   0.9  
#> 5            0.058          0.245            0.187             85   0.925
#> 6            0.045          0.258            0.214             90   0.95 
#> 7            0.024          0.279            0.255             95   0.975
#> 8            0.006          0.297            0.291             97.5 0.988
#> 9           -0.016          0.319            0.335             99   0.995
#> # … with 2 more variables: $`P-value` <dbl>, $`S-value (bits)` <dbl>

(z <- curve_table(intervalsdf[[1]], format = "latex"))

#> 
#> 
#> |     | Lower Limit| Upper Limit| Interval Width| Interval Level (%)|   CDF| P-value| S-value (bits)|
#> |:----|-----------:|-----------:|--------------:|------------------:|-----:|-------:|--------------:|
#> |2501 |       0.131|       0.172|          0.041|               25.0| 0.625|   0.750|          0.415|
#> |5001 |       0.108|       0.195|          0.088|               50.0| 0.750|   0.500|          1.000|
#> |7501 |       0.077|       0.226|          0.149|               75.0| 0.875|   0.250|          2.000|
#> |8001 |       0.068|       0.235|          0.166|               80.0| 0.900|   0.200|          2.322|
#> |8501 |       0.058|       0.245|          0.187|               85.0| 0.925|   0.150|          2.737|
#> |9001 |       0.045|       0.258|          0.214|               90.0| 0.950|   0.100|          3.322|
#> |9501 |       0.024|       0.279|          0.255|               95.0| 0.975|   0.050|          4.322|
#> |9751 |       0.006|       0.297|          0.291|               97.5| 0.988|   0.025|          5.322|
#> |9901 |      -0.016|       0.319|          0.335|               99.0| 0.995|   0.010|          6.644|